Curl::Multi for Solaris
Hi Keith,

Thanks for writing this—we thought we'd need to implement it all ourselves.

I did hit a couple snags when I went to deploy on Solaris, though. Here's what
I needed to change. It all could be a bit more elegant, but I tried to stay
within the established style. I tested the changes on OS X and Ubuntu, too.